Cleanup, moved general machinery out of test, wrote map function in Scala instead of elm
